+++ DataModels/Entities/LogBookTrainingDetail.cs
@@ -6,10 +6,10 @@ namespace DataModels.Entities
     {
         public long Id { get; set; }
         public long LogBookId { get; set; }
-        public Int16 DualGiven { get; set; }
-        public Int16 DualReceived { get; set; }
-        public Int16 SimulatedFlight { get; set; }
-        public Int16 GroundTraining { get; set; }
+        public double DualGiven { get; set; }
+        public double DualReceived { get; set; }
+        public double SimulatedFlight { get; set; }
+        public double GroundTraining { get; set; }
         public bool IsFlightReview { get; set; }
         public bool IsIPC { get; set; }
         public bool IsCheckRide { get; set; }
